+++ libs/validation/src/__tests__/app/vo/sample.dto.ts
@@ -7,6 +7,7 @@ import {
   MatchesWithProperty,
   PasswordLocalized,
   trimTransformer,
+  IsArrayCombinedLocalized,
 } from '../../../index';
 import { Optional } from '@nestjs/common';
 import { faker } from '@faker-js/faker';
@@ -79,6 +80,12 @@ export class SampleDto {
   @Optional()
   @IntegerType
   optionalNumber?: number;
+
+  @IsArrayCombinedLocalized({ minLength: 2, maxLength: 4 })
+  fileNameArray!: string[];
+
+  @IsArrayCombinedLocalized()
+  documentArray?: string[];
 }
 
 export const DEFAULT_SAMPLE_DTO: SampleDto = {
@@ -97,4 +104,9 @@ export const DEFAULT_SAMPLE_DTO: SampleDto = {
   middleName: faker.person.middleName(),
   someCheckboxValue: faker.datatype.boolean(),
   url: faker.internet.url(),
+  fileNameArray: [
+    `${faker.person.lastName()}.png`,
+    `${faker.person.lastName()}.png`,
+  ],
+  documentArray: [],
 };